# src/main/java/com/sh32bit/airline/response/ErrorResponse.java
package com.sh32bit.airline.response;

import lombok.AllArgsConstructor;
import lombok.Getter;
import lombok.Setter;

@Getter
@Setter
@AllArgsConstructor
public class ErrorResponse {
    private String message;
}